+++ tfe/data_source_organization_test.go
@@ -36,6 +36,32 @@ func TestAccTFEOrganizationDataSource_basic(t *testing.T) {
 	})
 }
 
+func TestAccTFEOrganizationDataSource_defaultProject(t *testing.T) {
+	skipUnlessBeta(t)
+
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+	org := &tfe.Organization{}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccTFEOrganizationDataSourceConfig_basic(rInt),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testAccCheckTFEOrganizationExists("tfe_organization.foo", org),
+					resource.TestCheckResourceAttrWith("tfe_organization.foo", "default_project_id", func(value string) error {
+						if value != org.DefaultProject.ID {
+							return fmt.Errorf("default project ID should be %s but was %s", org.DefaultProject.ID, value)
+						}
+						return nil
+					}),
+				),
+			},
+		},
+	})
+}
+
 func testAccTFEOrganizationDataSourceConfig_basic(rInt int) string {
 	return fmt.Sprintf(`
 resource "tfe_organization" "foo" {